use oninput instead of onkeyup and onchange
WARNING: this change doesn't work in IE 8 or below

pop_author and pop_biblio_rel used onkeyup and onchange. That
has the problem that when you use the mouse to click on an
author/biblio that you searched for, that click triggers the
onchange event and reloads the list which unselects the element
on which you clicked. oninput avoids that, because it is not triggered
by changes to the input field andn not by a mouseclick which deselects
the input field.

the rest of the code only used onkeyup. That has the problem
that it is not triggered by copy&paste if the paste is done
with the mouse