fix overdue days calculation when daylight savings time starts
the days between the due date and the real return date are calculated
by taking the unix-timestamp of each date, calculating the difference
and putting it in $_mksec. The days are then calculated by
intval($_mksec/(3600*24))
But if the daylight savings time changed inbetween, one of the days
has less than 3600*24 seconds. And because inval always rounds down,
the calculations counts one day fewer than it should.

My fix is to use round before the intval. The best solution would be
to use an existing library or PHP-functions for these calculations