Making toastr works inner colorBox popup
An example popup with colorbox is card setting or all about setting with form property action set to "submitExec" (iframe). Its using double iframe, First iframe in color box, and last is submitExec. Also stored toastr in $js and $css is not working. Although, it can be solved with change form instance to another with property action set to _self.