Add: en_AU help files
As per README included. These help files are an extended and enhanced version of the standard SLiMS 8 help files, written in Australian English, using terminology common to Australian libraries.
Spelling and grammar is standard Australian English.
My work on SLiMS translations is informed by the policies of the UN and the work of UNESCO.
https://www.unesco.org/en/languages-education/need-know

These files are released under the same license conditions as the SLiMS package that includes them.

Jim Richardson | jim@burmastudy.org | 06/10/2025